+++ VTK/Imaging/General/vtkImageGaussianSmooth.cxx
@@ -234,7 +234,7 @@ void vtkImageGaussianSmooth::ExecuteAxis(int axis, vtkImageData* inData, int inE
   int idxA, max;
   int wholeExtent[6], wholeMax, wholeMin;
   double* kernel;
-  // previousClip and currentClip rembers that the previous was not clipped
+  // previousClip and currentClip remembers that the previous was not clipped
   // keeps from recomputing kernels for center pixels.
   int kernelSize = 0;
   int kernelLeftClip, kernelRightClip;
@@ -303,8 +303,8 @@ void vtkImageGaussianSmooth::ExecuteAxis(int axis, vtkImageData* inData, int inE
     currentClipped = kernelLeftClip + kernelRightClip;
     if (currentClipped || previousClipped)
     {
-      this->ComputeKernel(kernel, -radius + kernelLeftClip, radius - kernelRightClip,
-        static_cast<double>(this->StandardDeviations[axis]));
+      this->ComputeKernel(
+        kernel, -radius + kernelLeftClip, radius - kernelRightClip, this->StandardDeviations[axis]);
       kernelSize = size - kernelLeftClip - kernelRightClip;
     }
     previousClipped = currentClipped;
